daemon: introduce the ability to modify the entrypoint
Certain distributions like Suse or RedHat have a downstream release
cadence different than upstream. As a result certain functionnalities
should be diseabled.
RHEL 7 is the first example of this, where we now have renamed the
entrypoint.sh to entrypoint.sh.in. In the Dockerfile of each distro that
don't want to customize the entrypoint we simply rename entrypoint.sh.in
to entrypoint.sh.
In distributions that want to customize the entrypoint we introduced
'TAG' in the entrypoint and a script to disable some scenario. So in
your Dockerfile you can simply call this script. Before doing that open
the 'disable_scenario' file and list the scenarii you want to disable.
During the build a new entrypoint.sh will be generated and the chosen
scenarii will be disabled.

Signed-off-by: Sébastien Han <seb@redhat.com>